CORE-8643 Refactor DND logic out of AppIntegrationPalette
